[feat] Implement MVVM pattern
